% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_calendar}
\alias{plot_calendar}
\title{Function to plot a commits calendar somehow like GitHub's}
\usage{
plot_calendar(df_commits, n_years = 3)
}
\arguments{
\item{df_commits}{data.frame(), a dataframe with a commit_date column}

\item{n_years}{number, represents how many years will be plotted}
}
\value{
a plotted chart
}
\description{
Function to plot a commits calendar somehow like GitHub's
}
\examples{
plot_calendar(df_commits)
plot_calendar(df_commits, n_years = 4)

}
\author{
Bruno Oliveira de Albuquerque, \email{bruno.albuquerque50@gmail.com}
}
